Moved Activity schemas to api_schemas and refactored code
