+++ noq_django/backend/serializers.py
@@ -8,6 +8,6 @@ class ActivitySerializer(serializers.ModelSerializer):
         model = Activity
         fields = ['id', 'title', 'description', 'start_time', 'end_time', 'is_registered']
 
-    def get_is_registered(self, obj):
+    def get_is_registered(self, obj) -> bool:
         registered_ids = self.context.get('registered_ids', [])
         return obj.id in registered_ids
